Calculate Time to Feature Cohort

This function calculates the time (in days) from the start date of a target cohort to the start date of a feature cohort for subjects in the specified cohorts. It generates a summary table of the time differences as well as a violin plot visualizing the distribution of these differences.

timeToFeatureCohort(
  targetCohortIds,
  featureCohortIds,
  targetCohortTableName,
  featureCohortTableName = targetCohortTableName,
  minDays = NULL,
  maxDays = NULL,
  tempEmulationationSchema = getOption("sqlRenderTempEmulationSchema")
)

Arguments

targetCohortIds

A vector of one or more Cohort Ids corresponding to target cohort(s). These records are collapsed to a unique combination of subject_id and cohort_start_date.

featureCohortIds

A vector of one or more Cohort Ids corresponding to feature cohort(s). These records are collapsed to a unique combination of subject_id and cohort_start_date.

targetCohortTableName

The name of the target cohort table.

featureCohortTableName

The name of the feature cohort table. Default is the same as targetCohortTableName.

minDays

(Optional) Minimum number of days between the target and feature cohort start dates. If provided, only records with a time difference greater than or equal to minDays are included.

maxDays

(Optional) Maximum number of days between the target and feature cohort start dates. If provided, only records with a time difference less than or equal to maxDays are included.

tempEmulationationSchema

Some database platforms (e.g., Oracle, Impala) do not support temporary tables natively. To emulate temp tables, provide a schema with write privileges where temporary tables can be created. Default is obtained from getOption("sqlRenderTempEmulationSchema").

Value

A list with two components:

summaryStatistics

A data frame containing summary statistics for the time differences (in days) grouped by the database identifier.

violinPlot

A violin plot (created by createViolinPlot) that visualizes the distribution of the time differences.

Details

The function executes a SQL query that:

Extracts unique subject start dates for the target cohort and assigns a sequence number for each subject.

Joins the target cohort with the feature cohort to find the first occurrence (minimum start date) in the feature cohort that occurs on or after the target start date.

Calculates the time difference (in days) between the target and feature cohort start dates.

Applies optional filtering based on minDays and maxDays.

The resulting data is then transformed into a long format (expanding counts by the number of subjects) and used to create both a violin plot and a table of summary statistics.

Examples

if (FALSE) { # \dontrun{
results <- timeToFeatureCohort(
  targetCohortIds = c(1, 2),
  featureCohortIds = c(3),
  targetCohortTableName = "target_cohort",
  featureCohortTableName = "feature_cohort",
  minDays = 30,
  maxDays = 365,
  tempEmulationationSchema = "temp_schema"
)

# Access the summary statistics
summaryStats <- results$summaryStatistics

# Display the violin plot
print(results$violinPlot)
} # }